% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proporciones.R
\name{plotpFG}
\alias{plotpFG}
\title{Grafico de las proporciones de grupos funcionales}
\usage{
plotpFG(proporcionFG)
}
\arguments{
\item{proporcionFG}{un data frame}
}
\value{
un gráfico de barras
}
\description{
Grafico de las proporciones de grupos funcionales
}
\examples{
\dontrun{
plotpFG(FunctionalGroup)}
}
