% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proporciones.R
\name{plotpHabitat}
\alias{plotpHabitat}
\title{Grafico de las proporciones de habitat}
\usage{
plotpHabitat(pH)
}
\arguments{
\item{pH}{un data frame}
}
\value{
un gráfico
}
\description{
Grafico de las proporciones de habitat
}
\examples{
plotpHabitat(Habitat)
}
