% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proporciones.R
\name{proporcionFG}
\alias{proporcionFG}
\title{Total de especies y proporciones por grupo funcional}
\usage{
proporcionFG(data)
}
\arguments{
\item{data}{un data frame}
}
\value{
una tabla con los siguientes campos:
        Name: nombre del ecosistema
        FunctionalGroup: categorias dentro de grupo funcional
        n: total de especies por categoria
        propFG: proporción de especies por categoria de grupo funcional
}
\description{
Total de especies y proporciones por grupo funcional
}
\examples{
FunctionalGroup <- proporcionFG(tabla)
}
