% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proporciones.R
\name{proporcionH}
\alias{proporcionH}
\title{Total de especies y proporciones por habitat}
\usage{
proporcionH(data)
}
\arguments{
\item{data}{un data frame}
}
\value{
una tabla con los siguientes campos:
        Name: nombre del ecosistema
        Habitat: categorias dentro de habitat
        n: total de especies por categoría
        propH: proporción de especies por categoria de habitat
}
\description{
Total de especies y proporciones por habitat
}
\examples{
Habitat <- proporcionH(tabla)
}
