% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proporciones.R
\name{readFile}
\alias{readFile}
\title{APERTURA ARCHIVO
Lee archivos que contengan el habitat y grupo funcional de las especies de uno o varios ecosistemas en formato .csv o separados por tabulaciones en columnas}
\usage{
readFile(fileName)
}
\arguments{
\item{fileName}{un archivo con el habitat y grupo funcional de cada especie}
}
\value{
un data frame
}
\description{
APERTURA ARCHIVO
Lee archivos que contengan el habitat y grupo funcional de las especies de uno o varios ecosistemas en formato .csv o separados por tabulaciones en columnas
}
\examples{

# Lee archivo con ecosistema, habitat y grupo funcional separado en columnas por tabulaciones

fileName <- system.file("exdata", "traits.txt", package = "Proporciones")
tabla <- readFile(fileName)
}
