% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion.R
\name{pH}
\alias{pH}
\title{CÁLCULO DEL TOTAL DE ESPECIES Y PROPORCIONES POR HÁBITATS Y GRUPOS FUNCIONALES}
\usage{
pH(data)
}
\arguments{
\item{data}{un data frame}
}
\value{
una tabla con los siguientes campos:
        Name: nombre del ecosistema
        Habitat/FunctionalGroup: categorias dentro de habitat/grupo funcional
        n: total de especies por categoría
        propH/propFG: proporción de especies por categoria
}
\description{
CÁLCULO DEL TOTAL DE ESPECIES Y PROPORCIONES POR HÁBITATS Y GRUPOS FUNCIONALES
}
\examples{

Habitat <- pH(tabla)
FunctionalGroup <- pFG(tabla)
PARA HÁBITATS
}
