% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion.R
\name{plotpHabitat}
\alias{plotpHabitat}
\title{GRÁFICO DE LAS PROPORCIONES DE HÁBITATS Y GRUPOS FUNCIONALES}
\usage{
plotpHabitat(pH)
}
\arguments{
\item{pH}{un data frame}

\item{pFG}{un data frame}
}
\value{
un gráfico
}
\description{
GRÁFICO DE LAS PROPORCIONES DE HÁBITATS Y GRUPOS FUNCIONALES
}
\examples{

plotpHabitat(Habitat)
plotpFG(FunctionalGroup)
PARA HÁBITATS
}
