% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proporciones.R
\name{plotpHabitat}
\alias{plotpHabitat}
\title{Grafico de las proporciones de habitat}
\usage{
plotpHabitat(proporcionH)
}
\arguments{
\item{proporcionH}{un data frame}
}
\value{
un gráfico de barras
}
\description{
Grafico de las proporciones de habitat
}
\examples{
\dontrun{
plotpHabitat(Habitat)}
}
